moved run method of assets writer to separate files
Moved the run method of the assets_writer_main to the separate assets_writer files.
Will extract a asset_writer class in the future to allow re-usability and more flexibility of the asset_writer.
